dds-4-25/requirements.txt
Django==1.6.2
Unidecode==0.04.14
amqp==1.4.3
anyjson==0.3.3
billiard==3.3.0.16
boto==2.25.0
celery==3.1.9
django-appconf==0.6
django-bootstrap-form==3.1
django-braces==1.3.1
django-celery==3.1.9
django-forms-bootstrap==3.0.0
django-jsonfield==0.9.12
django-model-utils==2.0.1
django-sendfile==0.3.4
django-stripe-payments==2.0b34
django-tastypie==0.11.0
django-user-accounts==1.0c1
eventlog==0.8.0
kombu==3.0.12
metron==1.3.1
pinax-theme-bootstrap==5.1.0
pymongo==2.6.3
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2013.9
requests==2.2.1
six==1.5.2
stripe==1.12.0
tablib==0.9.11
wsgiref==0.1.2
xlwt==0.7.5
